+++ apps/web/app/pages/signup.vue
@@ -1,5 +1,5 @@
 <template>
-  <EmailAuthForm
+  <AuthScreen
     mode="signup"
     title="auth.signupTitle"
     description="auth.signupDescription"